+++ src/constants/changelog.ts
@@ -2,6 +2,29 @@ import type { TReleaseNotes } from 'constants/types';
 
 const CHANGELOG: TReleaseNotes = [
   {
+    name: 'Version 4.9.5',
+    date: 1649374478404,
+    notes: [
+      {
+        description: 'Resetting encounters also resets badges now',
+        type: 'FIX',
+      },
+      {
+        description:
+          'Added Pewter City as a location in Red, Blue, and Yellow (requires a reset to encounters to see)',
+        type: 'FIX',
+      },
+      {
+        description: 'Added Zubat and Golbat to Iron Island encounters in BDSP',
+        type: 'FIX',
+      },
+      {
+        description: "Fixed Cranidos' level in Diamond, Pearl, and Platinum",
+        type: 'FIX',
+      },
+    ],
+  },
+  {
     name: 'Version 4.9.4',
     date: 1648869568970,
     notes: [